Refactor product compoennts
