+++ app/components/product/options.tsx
