+++ app/components/product/product-media.tsx
